Refine the Jira Query - 3 boards and 2 years
Refine the Jira query used to identify relevant issues for ingestion into the chatbot’s
vector database. This query will serve as the source of truth for the pilot run, ensuring
that the input prompts and expected outputs are based on the following boards:

- OSP
- OSPCIX
- OSPRH

The data is collected from the past two years.

OSPRH-15398